Switched from universal renderer to context and T proxy
This commit discards Nikhil Saraf's older architecture based on the SolidJS
universal renderer, in favor of his newer approach that looks more like Threlte:

https://github.com/nksaraf/react-three-fiber/blob/chess-solid-three/solid-example/src/components/ModernPiece.tsx

This approach uses context to maintain the scene graph, thus avoiding the need
for the universal renderer, which adds build-time complexity and a layer (or two)
of abstraction.  It also uses a "T" proxy object to instantiate the three.js objects
and add them to the scene appropriately.

Issues:

The example seems to leave the canvas at some default size, rather than filling the
viewport.

This approach still has a dependency on "scheduler", an NPM package apparently
created as a reusable React internal package.  We might consider replacing that.