remove applyProps from `core/utils` in favor of `solid/renderer`
Before commit there were two applyProps-implementations, one custom made and one from the original r3f-file.

This commit adjusts the custom applyProps function by removing the call to `useThree` and replacing it with `getRootState()`. This way that function can also in the code before the Provider and we can remove r3f's implementation.