optimization applyProps
Instead of `createRenderEffect(() => applyProps(instanceProps))` we now do `createRenderEffect(mapArray(() => Object.keys(instanceProps), ...)`.
This way we don't have to loop and re-create effects every single time a prop changes.

We also check if a prop has a getter or is a function. If they aren't we only set the property, otherwise we wrap it in an effect.