(fix) automatic resizing of canvas
r3f's code had a `store.subscribe(...)` in which they did a diff to check if they should update the camera. With solid we can just place them into a createEffect and let reactivity figure it out. Creates a flickering when resizing, idk if this is a regression or not.