(refactor) our code to look more like v9 branch
I copied the contents of r3f's `packages/fiber/src` and used the vscode-extension `Diff Folders` to compare the two directories and made changes so that these two directories look more alike.

Maybe this could be a way forward: have a r3f-directory which contains a clone of `packages/fiber/src` and we manually update them.