fiber: create tree with `props.children` instead of `getParent()`-context
Our previous architecture had a design flaw as described 023908c69b9639852e8fbe96f3cede1bd292acc8
```
const Noop = () => {
const box = <T.Mesh>...</T.Mesh>
return <></>
}
```
would add box to its parent, with or without returning the actual jsx.
This was due to the fact that parent-child relationships were established from the child to the parent with context through the `getParent()`-hook.

This commit changes the mechanism from child-to-parent to parent-to-child:
By walking the children from the parent and then attaching/connecting these childs, with a function called `parentChildren`.
This function is used when initializing scene too, the props.children are attached to the THREE.Scene with `parentChildren(() => instance, props)`

Since this would mean that props