fiber: update `inject` in Portal to be more in line with v8
Before this commit we were only setting a handful of parameters in inject and letting the rest be set by `rootState`.
This was causing bugs, p.ex in `GizmoHelperStories` from `drei`, where the camera of the `Portal` was being set by the `rootState`.
We were missing a couple of loc to filter out the correct keys to inject.

There is for sure a more performant, more solid-like way to merge the `rootState` and the `portalState`, which I will explore in a later commit.