fiber: `AllConstructorParameters`-typehelper to include overloaded ConstructorParameters
`ConstructorParameters` does not include overloaded constructors
see https://github.com/microsoft/TypeScript/issues/37079 and https://github.com/pmndrs/react-three-fiber/issues/581
`r3f`s solution is to special-case Color in v8 and solve it later for v9, because they faced some problems with the solution/hack described in the typescript github-issue
The hack didn't seem to cause the same issues with us, so I implemented that instead.

It's a manual union of overloaded methods, going up to 7 overloads, which is probably a bit of overkill.