fiber: seperate `IntrinsicElements` and `IntrinsicComponents`
Before commit we mirrored `r3f` `JSX.IntrinsicElements` with `SolidThree.IntrinsicElements`
`IntrinsicElements` is an interface with the props as values.

An earlier commit added `T.Suspense` and `T.Primitive` to the interface with `Primitive: Parameters<typeof Primitive<T>>[0]` but this didn't work nicely with the generic.
To solve this I added `IntrinsicComponents` to the `SolidThree`-namespace, `Primitive: typeof Primitive` and then in `SolidThreeElements = { ...SolidThree.IntrinsicElements } & { ... SolidThree.IntrinsicComponents