fiber: initial `inspect`-run
Introduce an initial `inspect`-run in which we resolve children and props, but do not yet create the three-element, attach/connect the children or apply the props.
This way any resource that is present in the new tree gets accessed and their respective `Suspense`s are activated.
The children are resolved depth-first and are then created from the leafs back to the root, until they are finally added to the rest of the scene.

see https://github.com/solidjs-community/solid-three/pull/12#issuecomment-1682258922 for a more detailed description of the timeline of the components.